requirements.txt
ajaxuploader==0.3.8
boto==2.24.0
coverage==3.7.1
Django==1.6.7
django-analytical==0.17.1
django-angular==0.7.2
django-annoying==0.7.6
django-autoslug==1.6.1
django-celery==3.0.23
django-compressor==1.3
django-debug-toolbar==1.0.1
django-extensions==1.3.3
django-localflavor==1.0
django-nose==1.2
django-pagination==1.0.7
django-passwords==0.2.0
django-redis-cache==0.11.1
django-ses==0.4.1
django-sslify==0.2
dj-database-url==0.2.1
djorm-ext-core==0.7
djorm-ext-expressions==0.6
-e git+https://github.com/revsys/django-tos.git@490c1edc674da4d1b7fb7203b034fa9b186cf848#egg=django_tos-master
-e git+https://github.com/niwibe/djorm-ext-pgjson@8fd70a13cb2597b56d1042539917784170569b40#egg=djorm_pgjson
-e git+https://github.com/gabrielfalcao/lettuce.git#egg=lettuce-master
-e git+https://github.com/buildingenergy/mcm-core.git@801f7b7ae34c32aa39f99ff567ac1e5ec25733cd#egg=mcm_core
-e hg+https://bitbucket.org/david/django-storages@e27c8b61ab57e5afaf21cccfee005c980d89480f#egg=django_storages
-e git+https://github.com/wellfire/django-organizations.git@f9da9778c212b253ebda49453cc57fe199020f51#egg=django_organizations
-e git+https://github.com/buildingenergy/python-dateutil.git@478b5799756a40a1f57a9a8b49a82f0c1e269165#egg=python-dateutil
-e git+https://github.com/buildingenergy/superperms.git@3d00d8ae14cd721fbef194067e2e30bc9acda96c#egg=superperms
nose-exclude==0.2.0
flake8==2.2.0
flower==0.6.0
fuzzywuzzy>=0.3.1
ipython==0.13.1
kombu==2.5.14
mock==1.0.1
newrelic==2.20.0.17
ngram==3.3.0
nose==1.3
pillow==2.4.0
psycopg2==2.5.1
python-memcached==1.53
pytz==2012h
PyYAML==3.10
raven==4.0.4
redis==2.9.1
regex==2014.01.10
requests==1.1.0
salad==0.4.14
simplejson==3.3.0
six==1.5.2
South==0.8.4
Sphinx==1.2.2
unicodecsv==0.9.4
xlwt==0.7.5
xmltodict==0.9.0
